refactor: move transaction builder to own file
