+++ packages/shared/src/lib/core/transactions/index.ts
@@ -1,1 +1,2 @@
 export * from './types'
+export * from './utils'
