+++ packages/shared/src/lib/core/transactions/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './buildPersistedEvmTransactionFromBlockscoutTransaction'
